Moved function docstring to types module to avoid circular import
The docstring function uses the _get_data_fields function from the types
module. To avoid a circular import we could import the types module
inside the docstring function, but I think it is better to just move the
docstring function to the types module, since it does not have any other
dependencies.